Use enable-test-discovery
Also split S3 tests between ones testing extension code and ones testing core library
