+++ .github/workflows/ci.yml
@@ -57,4 +57,4 @@ jobs:
         apt-get update -qq
         apt-get install -q -y tzdata zlib1g-dev
     - name: Test
-      run: swift test
+      run: swift test --enable-test-discovery --enable-test-discovery --sanitize=thread
